// Stock Pilot/stock-copilot-frontend-project-dev/stock-copilot-web-project/src/app/core/models/account-model.ts
import { BaseModel } from "../base/base-model";

export interface AccountModel extends BaseModel {
    address: string;
    branch: string;
    description: string;
    id: number;
    lastname: string;
    mail: string;
    name: string;
    password: string;
    phone: string;
    role: any;
    token: string;
}